Fix early card detection issue
If a card was detected before choosing the mode (with SAM, without
SAM...), the card was considered to be in the field even if it was
removed afterwards. This was an issue because selecting a mode was not
relaunching the card detection, resulting in a blocking state where the
application can not detect any tag.

The card detection is now launched when a mode is selected by the user
and not when the reader is ready.

Jira-Id: MOB-5163
Change-Id: I1137ddc7c2b888e1d0b80180dae1a042ecde2db6